update array batcher

+++ src/constbuffer_array_batcher.c
@@ -85,7 +85,7 @@ CONSTBUFFER_ARRAY_HANDLE constbuffer_array_batcher_batch(CONSTBUFFER_ARRAY_HANDL
                 }
                 else
                 {
-                    size_t all_buffers_array_end = all_buffers + malloc_size;
+                    size_t all_buffers_array_end = (size_t)(all_buffers + malloc_size);
                     uint32_t current_index = 0;
 
                     /* Codes_SRS_CONSTBUFFER_ARRAY_BATCHER_01_008: [ constbuffer_array_batcher_batch shall populate the first handle in the newly allocated handles array with the header buffer handle. ]*/
@@ -109,14 +109,14 @@ CONSTBUFFER_ARRAY_HANDLE constbuffer_array_batcher_batch(CONSTBUFFER_ARRAY_HANDL
 
                             (void)constbuffer_array_get_buffer_count(payloads[i], &buffer_count);
 
-                            for (j = 0; j < buffer_count && (all_buffers + current_index) < all_buffers_array_end; j++)
+                            for (j = 0; j < buffer_count && (size_t)(all_buffers + current_index) < all_buffers_array_end; j++)
                             {
                                 all_buffers[current_index++] = constbuffer_array_get_buffer(payloads[i], j);
                             }
                         }
 
                         result = constbuffer_array_create(all_buffers, all_buffers_array_size);
-                        for (i = 0; i < all_buffers_array_size && (all_buffers + i) < all_buffers_array_end; i++)
+                        for (i = 0; i < all_buffers_array_size && (size_t)(all_buffers + i) < all_buffers_array_end; i++)
                         {
                             CONSTBUFFER_DecRef(all_buffers[i]);
                         }
